requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
Pillow==10.2.0
moviepy==1.0.3
mutagen==1.47.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.112.0
pytz==2023.3
numpy==1.26.2
opencv-python==4.8.1.78
librosa==0.10.1
soundfile==0.12.1
scipy==1.11.4
matplotlib==3.8.2
